# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Meer API
  version: 0.1.0
  description: REST API for the Meer app (mobile). Uses JWT bearer auth.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /dashboard/users/{id}:
    get:
      summary: Get a user by id (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
...
  /stores:
    get:
      summary: List stores (paged)
      description: Public endpoint; provide Authorization for personalized fields, invalid tokens return 401.
      security: []
      parameters:
        - $ref: '#/components/parameters/XAppPackageHeader'
        - in: query
          name: type
          schema: { type: string, enum: [nearby] }
        - in: query
          name: categoryId
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 10 }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: lat
          schema: { type: number, format: double }
        - in: query
          name: lng
          schema: { type: number, format: double }
      responses:
        '200':
          description: Paged stores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseStoreCard'
    post:
      summary: Create store
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreCreateRequest'
      responses:
        '201':
          description: Created store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThriftStore'
        '400':
          description: Validation error
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /contents:
    get:
      summary: List guide contents
      description: Public endpoint; provide Authorization for likedByMe, invalid tokens return 401.
      security: []
      parameters:
        - $ref: '#/components/parameters/XAppPackageHeader'
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: sort
          schema:
            type: string
            enum: [newest, oldest]
            default: newest
        - in: query
          name: storeId
          required: false
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Paged guide content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseGuideTop'
    post:
      summary: Create guide content
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentCreateRequest'
      responses:
        '200':
          description: Created content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuideTop'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Store not found
  /contents/{id}:
    get:
      summary: Get content by id
      description: Public endpoint; provide Authorization for likedByMe, invalid tokens return 401.
      security: []
      parameters:
        - $ref: '#/components/parameters/XAppPackageHeader'
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Content detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuideTop'
        '404':
          description: Not found
    put:
      summary: Update content
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuideContentUpdateRequest'
      responses:
        '200':
          description: Updated content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuideTop'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Not found
    delete:
      summary: Delete content
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Not found
...
  /contents/{id}/comments:
    get:
      summary: List comments for content
      description: Public endpoint; provide Authorization for personalization, invalid tokens return 401.
      security: []
      parameters:
        - $ref: '#/components/parameters/XAppPackageHeader'
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paged comments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseGuideContentComment'
        '404':
          description: Content not found
    post:
      summary: Create comment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuideContentCommentRequest'
      responses:
        '200':
          description: Created comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuideContentComment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Content not found
  /contents/{id}/comments/{commentId}:
    patch:
      summary: Edit comment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: path
          name: commentId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuideContentCommentRequest'
      responses:
        '200':
          description: Updated comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuideContentComment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Comment not found
    delete:
      summary: Delete comment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: path
          name: commentId
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Comment not found
  /dashboard/stores:
    get:
      summary: Admin stores list
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
        - in: query
          name: search
          schema: { type: string }
          description: Case-insensitive search across store name, tagline, description, neighborhood
        - in: query
          name: sort
          schema:
            type: string
            enum: [newest, oldest]
            default: newest
      responses:
        '200':
          description: Paged stores (admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseDashboardStoreSummary'
  /dashboard/contents:
    get:
      summary: Admin contents list
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: sort
          schema:
            type: string
            enum: [newest, oldest]
            default: newest
      responses:
        '200':
          description: Paged contents (admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseGuideTop'
  /dashboard/contents/{id}:
    get:
      summary: Admin content detail
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Content detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuideTop'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    delete:
      summary: Delete content (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /dashboard/contents/{id}/restore:
    post:
      summary: Restore content (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Restored content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuideTop'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /dashboard/contents/{id}/comments:
    get:
      summary: Admin content comments list
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paged comments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseContentComment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /dashboard/contents/{id}/comments/{commentId}:
    delete:
      summary: Delete content comment (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: path
          name: commentId
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /dashboard/comments:
    get:
      summary: Admin comments queue
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
        - in: query
          name: sort
          schema:
            type: string
            enum: [newest, oldest]
            default: newest
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: contentId
          schema: { type: integer }
        - in: query
          name: storeId
          schema: { type: string, format: uuid }
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        '200':
          description: Paged comments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseContentComment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /dashboard/users:
    get:
      summary: Admin users list
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: sort
          schema:
            type: string
            enum: [newest, oldest]
            default: newest
      responses:
        '200':
          description: Paged users (admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseAdminUser'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XAppPackageHeader:
      name: X-App-Package
      in: header
      required: true
      schema:
        type: string
      description: Required for public GET endpoints (anonymous mode).
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
...
  schemas:
...
    Profile:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [USER, ADMIN] }
        avatarUrl: { type: string }
        bio: { type: string }
        notifyNewStores: { type: boolean }
        notifyPromos: { type: boolean }
        ownedThriftStore:
          $ref: '#/components/schemas/ThriftStore'
        createdAt: { type: string, format: date-time }
...
    ThriftStore:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        tagline: { type: string }
        coverImageUrl: { type: string }
        addressLine: { type: string }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
        openingHours: { type: string }
        facebook: { type: string }
        instagram: { type: string }
        website: { type: string }
        phone: { type: string }
        whatsapp: { type: string }
        categories:
          type: array
          items: { type: string }
        rating: { type: number, format: double }
        reviewCount: { type: integer }
        distanceMeters: { type: number, format: double }
        walkTimeMinutes: { type: integer }
        neighborhood: { type: string }
        badgeLabel: { type: string }
        isFavorite: { type: boolean }
        isOnlineStore: { type: boolean }
        description: { type: string }
        createdAt: { type: string, format: date-time }
        images:
          type: array
          items: { $ref: '#/components/schemas/StoreImage' }
...
    StoreCreateRequest:
      type: object
      required: [name, description, addressLine, phone, latitude, longitude]
      properties:
        name: { type: string, maxLength: 120 }
        description: { type: string, maxLength: 2000 }
        openingHours: { type: string, maxLength: 256 }
        addressLine: { type: string, maxLength: 512 }
        phone: { type: string, maxLength: 32 }
        email: { type: string, format: email, maxLength: 320 }
        neighborhood: { type: string, maxLength: 120 }
        tagline: { type: string, maxLength: 280 }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
        isOnlineStore: { type: boolean }
        categories:
          type: array
          maxItems: 10
          items: { type: string }
        social:
          $ref: '#/components/schemas/StoreSocialRequest'
...
    ContentCreateRequest:
      type: object
      required: [title, description]
      properties:
        title: { type: string, maxLength: 160 }
        description: { type: string, maxLength: 2000 }
        storeId:
          type: string
          format: uuid
          description: Optional for ADMIN users (creates a global post). Mandatory for regular Store Owners.
...
    PageResponseStoreCard:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ThriftStore'
        page: { type: integer }
        hasNext: { type: boolean }
...
    PageResponseGuideTop:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GuideTop'
        page: { type: integer }
        hasNext: { type: boolean }
...